requirements-dev.txt
-r requirements-ci.txt
pytest-sugar==0.9.0
mypy==0.530
